# frontend/app/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ConsentIn": {
        "properties": {
          "privacy_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Privacy Version",
            "type": "string"
          },
          "terms_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "terms_version",
          "privacy_version"
        ],
        "title": "ConsentIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Liveness": {
        "properties": {
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "Liveness",
        "type": "object"
      },
      "MeOut": {
        "properties": {
          "consent_required": {
            "title": "Consent Required",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "privacy_version": {
            "title": "Privacy Version",
            "type": "string"
          },
          "terms_version": {
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "consent_required",
          "terms_version",
          "privacy_version"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "Readiness": {
        "properties": {
          "checks": {
            "additionalProperties": {
              "enum": [
                "ok",
                "fail"
              ],
              "type": "string"
            },
            "title": "Checks",
            "type": "object"
          },
          "status": {
            "enum": [
              "ready",
              "not_ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "checks"
        ],
        "title": "Readiness",
        "type": "object"
      },
      "RevokeOthersOut": {
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "integer"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "RevokeOthersOut",
        "type": "object"
      },
      "SessionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "title": "Next Cursor",
            "type": "null"
          }
        },
        "required": [
          "items"
        ],
        "title": "SessionList",
        "type": "object"
      },
      "SessionOut": {
        "properties": {
          "absolute_expires_at": {
            "format": "date-time",
            "title": "Absolute Expires At",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "idle_expires_at": {
            "format": "date-time",
            "title": "Idle Expires At",
            "type": "string"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "is_current": {
            "title": "Is Current",
            "type": "boolean"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "id",
          "created_at",
          "last_seen_at",
          "idle_expires_at",
          "absolute_expires_at",
          "ip",
          "user_agent",
          "is_current"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SignupIn": {
        "properties": {
          "age_confirmed": {
            "const": true,
            "description": "The user confirms they are 18 or older",
            "title": "Age Confirmed",
            "type": "boolean"
          },
          "privacy_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Privacy Version",
            "type": "string"
          },
          "terms_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "age_confirmed",
          "terms_version",
          "privacy_version"
        ],
        "title": "SignupIn",
        "type": "object"
      },
      "SignupOut": {
        "properties": {
          "return_to": {
            "title": "Return To",
            "type": "string"
          }
        },
        "required": [
          "return_to"
        ],
        "title": "SignupOut",
        "type": "object"
      },
      "SignupProfileOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "privacy_version": {
            "title": "Privacy Version",
            "type": "string"
          },
          "terms_version": {
            "title": "Terms Version",
            "type": "string"
          }
        },
        "required": [
          "email",
          "display_name",
          "terms_version",
          "privacy_version"
        ],
        "title": "SignupProfileOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BuildOne API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Liveness"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "health"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readiness"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Readyz",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/auth/callback": {
      "get": {
        "operationId": "callback_v1_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Identity-provider callback (browser navigation)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "get": {
        "operationId": "login_v1_auth_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start sign-in (browser navigation)",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Sign out: delete this session",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/signup": {
      "get": {
        "operationId": "get_signup_v1_auth_signup_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "The pending sign-up awaiting 18+ and terms acceptance",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "complete_signup_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm 18+, accept the terms, and create the account",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/signup/decline": {
      "post": {
        "operationId": "decline_signup_v1_auth_signup_decline_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Under 18: abandon the sign-up; nothing is stored",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "get_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "The signed-in user (works while consent is pending)",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/consent": {
      "post": {
        "operationId": "accept_consent_v1_me_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept the current terms and privacy notice",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/sessions": {
      "get": {
        "operationId": "list_sessions_v1_me_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List my active sessions",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/sessions/revoke-others": {
      "post": {
        "operationId": "revoke_other_sessions_v1_me_sessions_revoke_others_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeOthersOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Revoke all my sessions except this one",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/me/sessions/{session_id}": {
      "delete": {
        "operationId": "revoke_session_v1_me_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke one of my sessions",
        "tags": [
          "me"
        ]
      }
    }
  }
}